Refactor/shop filter form to component (#69)
* refactor: convert ShopFilterForm from Island to standard component

- Replace useState-based Island component with standard HTML form
- Implement toggle functionality using <details> tag instead of JavaScript
- Reduce JavaScript bundle size by removing client-side state management
- Form now uses native browser form submission (method="get")



* SearchForm to components

---------
