+++ app/routes/shops/index.tsx
@@ -4,7 +4,7 @@ import { Container } from '../../components/Container'
 import { PageHeading } from '../../components/PageHeading'
 import { ShopListCard } from '../../components/ShopListCard'
 import type { Meta } from '../../types/meta'
-import { ShopFilterForm } from '../../islands/ShopFilterForm'
+import { ShopFilterForm } from '../../components/ShopFilterForm'
 import { buildShopFilterCondition } from '../../utils/buildShopFilterCondition'
 import { Pagination } from '../../components/Pagination'
 import { config } from '../../siteSettings'
